move classes from util to flap bearing because tomat did big dumb + refactor
