+++ common/src/main/kotlin/org/valkyrienskies/clockwork/ClockworkModClient.kt
@@ -4,7 +4,7 @@ import com.simibubi.create.foundation.outliner.Outliner
 import dev.architectury.event.events.client.ClientTickEvent
 import dev.architectury.event.events.common.TickEvent
 import org.valkyrienskies.clockwork.content.logistics.solid.delivery.frequency_slot.FrequencySlotGlobals
-import org.valkyrienskies.clockwork.util.render.DualLinkRenderer
+import org.valkyrienskies.clockwork.content.contraptions.flap.dual_link.DualLinkRenderer
 import org.valkyrienskies.kelvin.KelvinMod
 import org.valkyrienskies.kelvin.impl.client.DuctNetworkClient
