+++ common/src/main/kotlin/org/valkyrienskies/clockwork/content/contraptions/flap/FlapBearingBlockEntity.kt
@@ -6,7 +6,6 @@ import com.simibubi.create.content.contraptions.ControlledContraptionEntity
 import com.simibubi.create.content.contraptions.bearing.BearingBlock
 import com.simibubi.create.content.contraptions.bearing.IBearingBlockEntity
 import com.simibubi.create.content.kinetics.base.KineticBlockEntity
-import com.simibubi.create.content.redstone.link.LinkBehaviour
 import com.simibubi.create.foundation.blockEntity.behaviour.BlockEntityBehaviour
 import com.simibubi.create.foundation.blockEntity.behaviour.ValueBoxTransform
 import com.simibubi.create.foundation.utility.AngleHelper
@@ -21,7 +20,7 @@ import net.minecraft.world.level.block.entity.BlockEntityType
 import net.minecraft.world.level.block.state.BlockState
 import net.minecraft.world.level.block.state.properties.BlockStateProperties
 import org.apache.commons.lang3.tuple.Pair
-import org.valkyrienskies.clockwork.content.contraptions.flap.attached_frequency.SmartFlapBearingBehaviour
+import org.valkyrienskies.clockwork.content.contraptions.flap.dual_link.DualLinkBehaviour
 import org.valkyrienskies.clockwork.content.contraptions.flap.contraption.FlapContraption
 import org.valkyrienskies.clockwork.util.ClockworkConstants
 import org.valkyrienskies.clockwork.util.blocktype.ConnectedWingAlike
@@ -44,8 +43,8 @@ class FlapBearingBlockEntity(type: BlockEntityType<*>?, pos: BlockPos, state: Bl
     private var redstoneLevel = 0
     private var redstonePos: BlockPos? = null
 
-    private var linkFirst: SmartFlapBearingBehaviour? = null
-    private var linkSecond: SmartFlapBearingBehaviour? = null
+    private var linkFirst: DualLinkBehaviour? = null
+    private var linkSecond: DualLinkBehaviour? = null
 
     var firstDominant = true
 
@@ -88,8 +87,8 @@ class FlapBearingBlockEntity(type: BlockEntityType<*>?, pos: BlockPos, state: Bl
         val valueBoxes = ValueBoxTransform.Dual.makeSlots { first: Boolean -> FlapBearingFrequencySlot(first, true) }
         val valueBoxesSecond = ValueBoxTransform.Dual.makeSlots { first: Boolean -> FlapBearingFrequencySlot(first, false) }
 
-        linkFirst = SmartFlapBearingBehaviour(this, valueBoxes, {setFirstSignal(it)}, true)
-        linkSecond = SmartFlapBearingBehaviour(this, valueBoxesSecond, {setSecondSignal(it)},false)
+        linkFirst = DualLinkBehaviour(this, valueBoxes, {setFirstSignal(it)}, true)
+        linkSecond = DualLinkBehaviour(this, valueBoxesSecond, {setSecondSignal(it)},false)
     }
 
     fun setFirstSignal(power: Int) {
